app/pyproject.toml
[tool.poetry]
name = "url shortner"
version = "0.1.0"
description = "URL Shortner"
authors = ["Dilip Kumar Sharma <diliplakshya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"


[tool.poetry.group.core.dependencies]
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
cassandra-driver = "^3.28.0"
pydantic-settings = "^2.0.3"
gunicorn = "^21.2.0"
email-validator = "^2.0.0.post2"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
